+++ src/fftw/fftw_inplace.cc
@@ -30,7 +30,7 @@ void fftw_inplace(CTensor& in, int direction) {
 void fftw_inplace(CTensor& in, index dim, int direction) {
   assert(dim >= 0 && dim < in.rank());
   fftw_complex* pin = reinterpret_cast<fftw_complex*>(in.begin());
-  do_fftw(pin, pin, dim, in.dimensions(), direction);
+  do_fftw(pin, pin, (int)dim, in.dimensions(), direction);
 }
 
 void fftw_inplace(CTensor& in, const Booleans& convert, int direction) {
